requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
carvekit==4.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
diffusers @ git+https://github.com/huggingface/diffusers.git@151b74cd7758df590c523230a86230ba3bbc786f
DistVAE==0.0.0b4
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.13.1
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
gradio==5.0.2
gradio_client==1.4.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
iniconfig==2.0.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
-e git+https://github.com/luca-medeiros/lang-segment-anything.git@19cb3670caea21e11b35e01ce1115a4a35b27fb5#egg=lang_sam
litserve==0.2.3
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
ninja==1.11.1.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum-quanto==0.2.6
orjson==3.10.12
packaging==24.2
pandas==2.2.3
peft==0.14.0
piexif==1.1.3
pillow==11.1.0
pluggy==1.5.0
portalocker==3.0.0
protobuf==5.29.2
psutil==6.1.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
ray==2.40.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.1
safehttpx==0.1.1
safetensors==0.4.5
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2@c2ec8e14a185632b0a5d8b161928ceb50197eddc
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
supervision==0.23.0
sympy==1.13.1
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.5.1+cu124
torchao==0.8.0
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.44.2
triton==3.1.0
typer==0.14.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.20.0
watchfiles==1.0.3
websockets==12.0
Werkzeug==3.1.3
xfuser==0.4.1
yunchang==0.6.0
pillow==11.1.0
transformers==4.44.2
accelerate==1.2.1
python-multipart==0.0.12
uvicorn==0.32.1
diffusers
protobuf
aiofiles==23.2.1
fastapi==0.115.5
filelock==3.13.1
huggingface-hub==0.26.3
torch
sentencepiece
